[IMP] planning: click on material resource shouldn't possible
In This commit,  we restrict clicking on material resources as the resource's
form view does not offer important information. Therefore, it may not be very
engaging to present it prominently. Additionally, it remains accessible through
the internal link. In general, we tend to restrict the number of clickable
elements in the list view.

task-3277195

Part-of: odoo/enterprise#39926