[IMP] planning: make material resource future shifts open when archive
In This commit,  when a material resource is archived, we transform all of its
future shifts into open shifts to be consistent with the behavior we have for
human resources. If a material resource is archived, it is not expected to be
used anymore. It is thus misleading to leave shifts assigned to this resource.

task-3277195

closes odoo/enterprise#39926

Related: odoo/odoo#125533
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>