[IMP] knowledge: open article in new tab outside of Knowledge
When clicking on an article link outside of Knowledge, it should open a new tab
to start a new navigation, so that one can still edit/consult the record with
the link and do a documentation search in Knowledge on the side.

Attributes on the anchor for a Behavior are now "fixed" in a separate method:
`setupAnchor`. They can be "wrong" for various reasons:
- cleaned by sanitization (frontend, backend). Indeed, it is not necessary to
  add an exception during those sanitizations for attributes whose value depend
  on the state of the application
- wrong attribute value from a previous Odoo version
- wrong attribute from content pasted/dropped from an origin field in another
  state than the destination field (i.e. copy/paste from a 'readonly' field to
  an 'editable' field.

Task-2852931

closes odoo/enterprise#39953

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>