[FIX] planning: fix traceback in number of occurrences
Description of the issue/feature this PR addresses:
In the planning view, to create a recurring shift, please ensure to select the
repeat field and choose repeat_type to  the "Number of Repetitions" option.
Next, input the repeat_number  field. Please note that this field can
accommodate more than 6 digits,

Current behavior before PR:
After clicking on the save button, a traceback error is being displayed.

Desired behavior after PR is merged:
it should not be showing traceback

fix:
When the repeat_unit is set to 'year,' the repeat_qty is subject to a maximum
limit of 7976. To ensure accurate data processing, a validation message should
be displayed to the user if the input value exceeds a certain number of digits.

task-3280609

closes odoo/enterprise#39976

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>