[FIX] ln10_be_hr_contract_salary: check if salary advantage exists in contract form
Before, if the private car reimbursed amount advantage was archived, the input
would not be displayed in the contract web form. However, `onchangeCompanyCar`
and `onchangePrivateCar` both expected the other field to always exist.

This would throw a: `Cannot read properties of undefined (reading 'checked')`
when trying to evaluate `private_car_input[0].checked`.

The solution is to make sure that the jquery request returns something before
evaluating.

opw-3236066

closes odoo/enterprise#40044

X-original-commit: 0004999dacc8a719baf687c4ed9a1f3a5f91c511
Signed-off-by: Kevin Baptiste <kba@odoo.com>