[FIX] sale_subscription: prevent traceback when first_contract_date
is not defined on upsells.

first_contract_date is not always properly defined on upsells. As the origin_order_id
is correct, we decide to rely on origin_order_id to determine the first contract date

first_contract_date field should only be used for reporting purpose.

This commit fixes the following traceback:

```py
  File "/home/odoo/src/custom/private/openerp_enterprise/models/sale_subscription.py", line 1471, in _prepare_invoice_line
    res = super()._prepare_invoice_line(**optional_values)
  File "/home/odoo/src/odoo/saas-16.2/addons/sale_project/models/sale_order_line.py", line 282, in _prepare_invoice_line
    values = super(SaleOrderLine, self)._prepare_invoice_line(**optional_values)
  File "/home/odoo/src/enterprise/saas-16.2/sale_subscription/models/sale_order_line.py", line 251, in _prepare_invoice_line
    new_period_start = max(self.order_id.start_date or fields.Datetime.today(), self.order_id.first_contract_date)
TypeError: '>' not supported between instances of 'bool' and 'datetime.date'
```

Part-of: odoo/enterprise#40461