[FIX] sale_amazon: force save company_id on amazon account
To reproduce (V16.2):

 1. Install sale_amazon without demo datas
(otherwise you will have several companies and the issue won't reproduce)
 2. Try to create an amazon account
 3. Python Traceback:
 ```
 ...
 File "odoo/enterprise/sale_amazon/models/amazon_account.py", line 225, in create
    'location_id': parent_location_data[0]['view_location_id'][0],
IndexError: list index out of range
 ```

This happen as the `company_id` field is missing from the JS `create` request.
The issue happen as the `company_id` is loaded only when the user is in the group
`base.group_multi_company` (due to JS performance improvements in 16.2).
This is for this reason that the issue won't happen in multi-company.

opw-3268920

closes odoo/enterprise#40070

X-original-commit: c0e2022ff400edbaa35e98a0b035111989fcc609
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>
Signed-off-by: Sens Loan (lse) <lse@odoo.com>