[FIX] web_studio: correctly remove fields with groups from activeFields
Commit odoo/odoo@baebb6a5b05ac8d59501a6071a5513e31a4ca047 introduced a universal way to handle
multiple instances of the same field in a view arch.
To do so, it separates what fields are to be fetched (activeFields) from the "instances" of those
fields in the arch (fieldNodes). Studio's code was only partially ready, relying on the fact
that the key for fieldNode was the field's name, which isn't true any longer.

This commit fixes that, and removes from the fieldNodes the ones that have studio_no_fetch
and more importantly removes them from activeFields.

To reproduce: have a field with a group declared in python, which the user doesn't have.
Open studio with a view that has that field in the arch.

Before this commit, there was an access rights crash.

After this commit, there is no crash.

closes odoo/enterprise#39997

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>