[REF] documents: rework multi edit logic
This commit reworks the multi edit logic of the list and kanban
views of documents, following the refactoring of odoo/odoo#119082.

Instead of hacking the update method of Record, the model of those
views are now multi editable. This allows to fix non blocking small
issues: documents are removed from the view after editing their
workspace and in the kanban view, the star on records is now multi
editable as well (whereas before, it was, but only sometimes).

Part of task 3179751

closes odoo/enterprise#40012

Signed-off-by: Géry Debongnie <ged@odoo.com>