[FIX] stock_barcode: backorder link
When one or multiple (with picking batch) Back Orders are created from a
Barcode operation, a link to the created Back Order(s) is added to the
success notification.
This link opened an action thanks to the URL, but by doing that, the
action manager loses the breadcrumb.

This commit replaces the BO's link by a button which will call an
action to open the newly created BO when clicked.

Part-of: odoo/enterprise#38096