[FIX] helpdesk_stock: remove suitable sale orders for returns w/o ticket_id
Before this commit:
`suitable_sale_order_ids` is populated with all `done` pickings when
returning a transfer without a `ticket_id`, which crashes the frontend
upon loading.

Steps to reproduce:
    1. Install Stock & Helpdesk
    2. Create thousands of sale orders and mark them as done
    3. Return a transfer

After this commit:
Similarly to how `suitable_picking_ids` was fixed in [e764792],
set `suitable_sale_order_ids` to False when the transfer doesn't have
a `ticket_id`, as the fields are not used in that case anyway.

opw-3277004

[e764792]: https://github.com/odoo/enterprise/commit/e7647925addaad81891745d9e06e86bd8965883b

closes odoo/enterprise#40059

X-original-commit: fbdd6dfed8b8a3a374d075539e1ca7ab7cafe9e0
Signed-off-by: William Henrotin (whe) <whe@odoo.com>