[FIX] web_grid: click on cell to enter in edit mode in mobile
Before this commit, in mobile the user has to click on the cell to
focus it (mouse over event is triggered) and then click again to
enter in edit mode.

This commit adds a click event in the cell to directly enter in edit
mode in one click in mobile.

part of task-3226579
task-3283082

X-original-commit: 2b7e8b8b597b29ed7facb1fb3d13e1e67e9f5fcb
Part-of: odoo/enterprise#40460