[FIX] l10n_co_edi: Remove deprecated tags
Based on the lates DIAN specification, the tags <ADQ_8> and <ADQ_9> have been
removed, and now are required only for the EDI specific for the Health care
secto, which is not supported in Odoo as it requires additional features.
These fields were used to send the first name and last name of the customer when
they were clasified as "Individual", now the Customer name either for Individuals
and for companies is sent in the ADQ_6 and ADQ_7.

closes odoo/enterprise#40290

X-original-commit: 9351a4d243ee2fddb76ee14da8c120ad4ef53ad9
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Bhavin Patel (bat) <bat@odoo.com>