[IMP] sale_amazon: synchronize Odoo stock with Amazon
Currently, the stock on Amazon for FBM is not updated when it changes on odoo side. This can lead
to overselling, so the user either will have to cancel a lot of orders or he risks having big delay
to deliver. It can then have a bad impact on the reputation.

From now on, Odoo will regularly send the current available quantity of each offers for account
that activated the synchronization feature.

This only works for FBM products, as Amazon will not update the quantity for FBA even when we try.

task-2864098

closes odoo/enterprise#40720

Related: odoo/documentation#4773
Related: odoo/upgrade#4810
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>