[IMP] spreadsheet_edition: Introducing Version History
This commits adds the possibility for the users to consult the whole
edition history of their spreadsheets (which therefore includes
dashboards and templates as well).

Upon consultation, the users will be able to jump back to a previous
state of the spreadsheet and to restore that state in the form of a new
copy of the spreadsheet.

They will also have the possibility to name/rename the revisions
(i.e. modification steps) to organise their history at their will.

closes odoo/enterprise#40066

Task: 3142006
Related: odoo/upgrade#5229
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>