[FIX] stock_barcode: add CSS adaptations
Following the Kanban view refactor that came with WOWL, the `style`
attribute set on a button put in a kanban view is no longer used.
So instead we make use of the Bootstrap classes to set the paddings of
the extra button in the kanban view, since classes are properly handled.
This avoids breaking the layout of the header line.

Also removes the opacity dimming on the edit button of a picking, since
it kinda gave the impression that the button was disabled.

Part of task-3218314

closes odoo/enterprise#40153

Related: odoo/odoo#119381
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>