[FIX] hr_payroll_account: fix a failing condition in get_existing_lines
In `_get_existing_lines(...)` we check if line['analytic_distribution'] is set
along with two variables, if one of the three is set we do a subcheck. But if
only line['analytic_distribution'] wasn't set it will try to do something like
`variable in False` which raises a `bool is not iterable`error.

closes odoo/enterprise#40226

Task: 3283253
X-original-commit: e8c227691e6ca40b542cfcc7ea4d43f2cd5f71ab
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Claire Bretton <clbr@odoo.com>