[FIX] account_reports: fix display of outstanding payments
Problem: Currently, the outstanding payments only show on the bank reconciliation report if there are payments attached to journal entries.
However, outstanding payments refer to journal entries that use an outstanding account.
Thus, it is not required for the journal entries to have a payment in order for it to display on the bank reconciliation report.
By query, the journal entries using an outstanding payment account is correctly fetched, but not properly displayed.

Solution:The dictionary, grouped_results, should be grouped by the move_id instead of payment_id.
Move_id are guaranteed to have a unique value whereas payment_id is not since it can be false.
With this fix, the journal entries that uses an outstanding account but has no payment will be displayed.

opw-3230064

closes odoo/enterprise#40093

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>