[FIX] appointment: check appointment validity with big UTC offset
Steps to reproduce:
-------------------
- change the user's time zone to "Pacific/Auckland" (UTC +12);
- go to Online Appointments;
- create a slot on Monday from 9 am to 12 pm;
- publish it;
- go to the share link of the appointment type;
- try to choose an appointment for 9 am;

Remark: Set the browser timezone to "Pacific/Auckland".

Issue:
------
The request will return a 404 error.

Cause:
------
When comparing the slot chosen by the user and the slots
available for the appointment's staff user, we use the `weekday`.
https://github.com/odoo/enterprise/blob/e97419a617f22fc71042061e5c0e0e1d96d50e43/appointment/models/calendar_appointment_type.py#L199-L205
Unfortunately, with a very large UTC offset, in some cases
we change the day.
The result is that we do not find any slots.

Solution:
---------
Make sure that `first_day` and `last_day` are
in the timezone of the appointment.

opw-3267279

closes odoo/enterprise#41528

X-original-commit: ad1400e69a3c4cbbd433d2ba1c522317bbad9999
Signed-off-by: Lefebvre Thomas (thle) <thle@odoo.com>
Signed-off-by: Hennecart Jérémy (jeh) <jeh@odoo.com>