[FIX] account_reports: don't sort the loaded lines before applying the load_more_limit when expanding a groupby
To reproduce the issue:

1) Create a report with a line doing a group by partner_id.
2) Set the load_more_limit of this report to 2
3) Create 3 partners, named A, B and Z.
4) Create a move line matching the report line with partner Z, B and then A (in this order)
5) Open the report, and unfold the groupby line: only two partners should be loaded, and the 'load more' button shown.
        => They should be [B, Z], but instead they are [A, B].

This is because the groupby lines are displayed following the _order of the related model. Here, res.partner has _order='name'.
For the load_more feature, we need to load one more result than the limit allows (because if there is one more, we know the "Load More" button will need to be displayed).
The bug is here that the reordering of the result to match the model's _order is done before applying the load_more_limit. So, in our example, we retrieve [Z, B, A] from the database, then reorder it to [A, B, Z], and finally only keep the first 2 elements, so [A, B].

With this fix, we'll instead retrieve [Z, B, A], then keep A appart and reorder the first 2 elements to get [B, Z], re-add A to get [B, Z, A] in _expand_groupby, and finally remove the last element in _report_expand_unfoldable_line_with_groupby so that the final result if [B, Z].

X-original-commit: 73af7843949170dd88b0be6d2d07e11a1a69d669
Part-of: odoo/enterprise#40459