[IMP] sale_subscription: allow manage payment form for child partners
before this commit, managing the payment form was only possible if the
connected partner was the partner of the contract. Sometimes, the
contract is for the company and an employee (child partner) must be allowed
to manage the contract with his own login.

This commmit enable the manage payment form when the partner is the child.

X-original-commit: 7d357dc86a243e747d97c98b0cc6d233d7966e3d
Part-of: odoo/enterprise#40836