[REF] account_{accountant,followup,reports}: remove manual reonciliation widget
The manual reconciliation widget is too complex. We completely remove it and
replace it with a simple wizard `account.reconcile.wizard` that opens when
selected lines can't be directly/silently reconciled (then we need a
write-off whose data will be filled in the wizard).

All view that use to lead to the previous widget now lead to the `account.move.line`
list view filtred on posted and unreconciled moves.

We now use of the _reconcile_plan instead of the previous reconciliation
which should be more efficient and correct.

closes odoo/enterprise#40510

Task: 3060792
Related: odoo/upgrade#4674
Related: odoo/odoo#120832
Signed-off-by: Laurent Smet <las@odoo.com>