[IMP] documents_spreadsheet: include spreadsheets in sharing zip
In the Document app, if you select multiple documents and download them
(as a zip), spreadsheet are not downloadable. The reason is we can't export
them to an excel file server-side.

However, when a spreadsheet is shared, the excel file is saved along with the
spreadsheet copy. That means we can include this excel file when downloading
the documents zip from the portal view.

Spec
====

Create a sharing link with multiple documents, including a spreadsheet.
Open the sharing portal page and click on "Download all", the spreadsheet
is now part of the zip, as an excel file.

Task 3045808

closes odoo/enterprise#37687

Related: odoo/odoo#114040
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>