[FIX] account_accountant: Bank Recon table width and UI
The lines_widget on the bank recon width seems to be broken from saas-16.1 up.
This commit removes a style that was applied for 16.0 but is no longer required.

The manual operations tab also causes a small horizontal scroll on the bank recon form view. This is caused by the bootstrap row. Add css to remove the right & left padding.

Fix the lines_widget table header, to match the standard table header.

The notes field stays as is, and need not be fwd-ported

closes odoo/enterprise#40392

X-original-commit: 9b26e2bc7f836ef38886c8f2184169b93a9b636d
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Ayob Habib (ayh) <ayh@odoo.com>