[FIX] marketing_automation: use the right getter..
on default_dict in _compute_participants

This solves an issue with _compute_participants not setting any value
if a campaign wasn't present in the return of a _read_group.

We recently started using a defaultdict when adapting to the improved
_read_group method at 0a24c01ec178984df032710b81d562366c016ebb

defaultdict will not add new entries when using the `get` method.
It will instead return None like a regular dict.

We just replace .get with directly addressing the element.

task-3284558

closes odoo/enterprise#40151

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>