[FIX] web_grid: update in local before calling grid_update_cell
Before this commit, since the value in the cell was updated at the
end of the rpc call (grid_update_cell), a small ficker could appear
by seeing the old value and then the one selected by the user.

This commit directly update the cell before calling the rpc to
avoid having any flicker.

part of task-3226579
task-3283082

closes odoo/enterprise#40460

X-original-commit: 4a5dfb3e2b04ca4b652589b428a0ff5147179a2f
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>