[FIX] knowledge: make stars in list views yellow
Purpose:
--------
Currently, the star icon of the toggle_favorite button in the knowledge
list views is green, while it is yellow in every other list view. This is
because we do not use the `boolean_favorite_field` field in knowledge but
a button with an action so that users can add to or remove from their
favorites articles for which they only have read access.

This commit adds the same scss style used in the boolean favorite field
to the toggle favorite button to make the star yellow in knowledge's
list views (search, trashed and items).

Task-3284518

closes odoo/enterprise#40741

X-original-commit: 46aaa25bfb5a73b9014db2a7399b9624cab9cfaa
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>