[FIX] account_reports: fix average creditors days
Currently, the formula for calculating the average
creditors days in the executive summary is wrong.

It is using the operating income as the denominator,
instead of the the cost of goods sold, which can be
computed as the sum of the cost of sales and the
expenses (excluding depreciations).

opw-3222694

closes odoo/enterprise#40337

X-original-commit: d3e7976032f1616f0b93f820b11c5f86f8f7e935
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>