[FIX] account_reports: fix average creditors days
Currently, the formula for calculating the average
creditors days in the executive summary is wrong.

It is using the operating income as the denominator,
instead of the the cost of goods sold, which can be
computed as the sum of the cost of sales and the
expenses (excluding depreciations).

opw-3222694

closes odoo/enterprise#40271

X-original-commit: c351e88b7528cae8e652c5b096401e7eaad31ffa
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>