[FIX] l10n_cl_edi, _stock: Fix discount rounding in DTE xmls
According to the XSD definition `DTE_v10.xsd`, the Discount Percentage
field ('DescuentoPct') of invoice lines should be rounded to 2 decimal
places.

This should be the case both in the invoice DTE and in the DTE of the
Guía de Despacho.

With this commit, we use float_repr to truncate this field to 2 decimal
places.

opw-3202045

closes odoo/enterprise#40309

X-original-commit: 6c5417069ff6a278a8a15b0fb23867e0f6291c3f
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>