[FIX] account_intrastat: unwanted report lines in anglo-saxon
When using anglo-saxon accounting and automated inventory evaluation for
a product category, some move lines are generated on the move upon
posting. These "COGS" lines are unintentionally included in the intrastat
report from v16.0 onwards.

Before v16.0 a condition existed in the where clause of the query that
gathers the lines to represent in the intrastat report that excluded
lines that are not represented on the invoice tab.

The condition was previously `NOT move_line.exclude_from_invoice_tab`,
but was replaced with `move_line.display_type = 'product'`
https://github.com/odoo/enterprise/commit/7a237e665a179c6bd0d7962d331c10dbfe50c0b8#diff-7b1b123893744b02df76c7fa971ebb68b9f59e26f619e3a7d8f290c0f45ed037L210-L220

and then subsequently unintentially removed in reportalypse
https://github.com/odoo/enterprise/commit/f435654f70f3d6a1fec7fa696945f2c0eb9a7056#diff-7b1b123893744b02df76c7fa971ebb68b9f59e26f619e3a7d8f290c0f45ed037L215

This commit re-introduces this condition to the where-clause from v16.0 onwards

opw-3222394

closes odoo/enterprise#40191

Signed-off-by: Quentin De Paoli <qdp@odoo.com>