[FIX] account_accountant: bank balance confusion
Due to performance issues, the computation of the balance in GL was changed (odoo/odoo#103697)
Since it is computed in `_get_journal_dashboard_bank_running_balance` as the last statement balance (balance_end_real) + transactions (account.bank.statement.line since the last statement), we rename the label on the journal dashboard to "Running Balance".
Also, fix the order clause of the "Last Statement" balance as it has a problem when there are 2 statements on the same day.

OPW-3265511

closes odoo/enterprise#40334

X-original-commit: ff65d77fdacc5261aafd1e7e3ad4f7bc4ad3730a
Related: odoo/odoo#119764
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Ayob Habib (ayh) <ayh@odoo.com>