[FIX] account_reports: allow switching between variants when opening a localized tax report from the tax closing next activity
When opening a localized tax report (i.e. not the generic one) from the tax closing next activity (on the Accounting dashboard, after creating the closing move), it was impossible to switch between variants (selecting them had no effect at all). This was due to a javascript bug: the widget needs to keep track of the root report triggering the opening, and to get it, it took it from the options passed to the action, or (if not set) from the action's context. The usual opening flow of the reports (through the menus) does not pass any options in the action dict; however, it gets passed when going through the next activity. The bug was here caused by the fact the localized report was then used as root report, because it was in report_options['report_id']. The fix is here to rather take the report_id in the context of the action if it is provided, and fallback to previous_options['report_id'] only if it isn't set.

OPW 3274875

closes odoo/enterprise#40200

Signed-off-by: John Laterre (jol) <jol@odoo.com>