[FIX] *: adapt front-end HTML field sanitization
*: appointment, helpdesk, sign, website_twitter_wall

HTML fields that appear in the front-end can be modified using the
website editor. Some of them are sanitized in a way that breaks the
behavior of snippets that can be dropped within them.

This commit adapts the sanitization of those HTML fields so that the
snippets behave as expected.

Steps to reproduce:
- Install `website` and `sale_subscription`.
- Edit a Subscription.
- Enter "Test" in the Terms and Conditions field.
- Save.
- Go to Customer Preview.
- Edit and drop a Carousel snippet inside the Terms and Conditions.
- Save.

=> The Carousel buttons do not respond.

opw-3267589

closes odoo/enterprise#42802

X-original-commit: 2ce9908bdeb06e8de6b2b0187c6acccbc16ef2a0
Related: odoo/odoo#125650
Signed-off-by: Romain Derie (rde) <rde@odoo.com>