[FIX] *: adapt front-end HTML field sanitization
*: appointment, helpdesk, sign, website_twitter_wall

HTML fields that appear in the front-end can be modified using the
website editor. Some of them are sanitized in a way that breaks the
behavior of snippets that can be dropped within them.

This commit adapts the sanitization of those HTML fields so that the
snippets behave as expected.

Steps to reproduce:
- Install `website` and `sale_subscription`.
- Edit a Subscription.
- Enter "Test" in the Terms and Conditions field.
- Save.
- Go to Customer Preview.
- Edit and drop a Carousel snippet inside the Terms and Conditions.
- Save.

=> The Carousel buttons do not respond.

opw-3267589

closes odoo/enterprise#43311

X-original-commit: e65fe2d177ca8d0b6c8da8978560ff82a46e1f0f
Related: odoo/odoo#126708
Signed-off-by: Romain Derie (rde) <rde@odoo.com>
Signed-off-by: Benoit Socias (bso) <bso@odoo.com>