[FIX] account_reports : add tooltip to button box
Steps to reproduce the issue:

1. Go to any of the contacts with partner ledger option.
2. Make the browser window small and observe the text in the partner ledger button box.
3. The button will not have any tooltip

Current Behaviour:
The text in the button box does not get ellipsized, rather stays the same and may get out of the button box if the box becomes smaller and does not provide any tooltip.

Desired Behaviour:
The text in the button should be ellipsized and a tooltip should appear when hovered over the button.

OPW-3266235

closes odoo/enterprise#40948

X-original-commit: c7778322ffcb8a60c3dbe7e7fec6af4241adb0e6
Related: odoo/odoo#121167
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans (app) <app@odoo.com>