[FIX] account_reports : add tooltip to button box
Steps to reproduce the issue:

1. Go to any of the contacts with partner ledger option.
2. Make the browser window small and observe the text in the partner ledger button box.
3. The button will not have any tooltip

Current Behaviour:
The text in the button box does not get ellipsized, rather stays the same and may get out of the button box if the box becomes smaller and does not provide any tooltip.

Desired Behaviour:
The text in the button should be ellipsized and a tooltip should appear when hovered over the button.

OPW-3266235

closes odoo/enterprise#40224

Related: odoo/odoo#119531
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>