[FIX] l10n_lu_reports: fix mandatory fields
odoo/enterprise#26401 merged modules in a way that only added mandatory fields in the VAT declaration if it was an annual declaration.
They should be also added for monthly, quarterly, and simplified declarations.

At the same time, the test for the tax report was modified during the reportalypse and did not test the same xml file as the one normally generated for the user,
which is why this error was not caught in the test.
The xml file in the test contained fields that should be removed if they amount to zero and did not contain the required fields.
Now the test tests the same xml file a user would see when exporting the tax report.

opw-3270105

closes odoo/enterprise#40263

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>