[REF] iot,pos_iot,*: minimize the use of legacy web modules
*: account_accountant,pos_l10n_se,pos_restaurant_iot,pos_settle_due

This is part of the continuous effort of refactoring pos addons towards using
more modern modules and programming patterns (such as services). After this
commit, point_of_sale addons are now left with the use of the legacy
`web.concurrency` module because of the `MutexedDropPrevious`. It's okay to keep
because it's relatively an independent module compared to other legacy web
module.

See https://github.com/odoo/odoo/pull/117231 for more info.

closes odoo/enterprise#39078

Related: odoo/odoo#117231
Signed-off-by: Samuel Degueldre <sad@odoo.com>