[FIX] web: open mobile search view in a dialog
Before this commit, the mobile search view was sent to the end of the
dom (t-portal) and so was on top of everything even the dialog that
was opened after it. Now, the mobile search view is open in a dialog
and the depth level between it and other dialogs is fixed.

task id: 3273035

closes odoo/enterprise#40255

Related: odoo/odoo#119595
Signed-off-by: Mathieu Duckerts-Antoine <dam@odoo.com>