[IMP] l10n_pe_edi: change default IAP endpoints url
We used an old odoo.sh server for the test env
of the IAP part of l10n_pe_edi, but that subscription
always expires and it would be better to use
the regular IAP test env.

For that, we just need to change the URL in the code and
in the meantime, we also change the production url
for consistency.

closes odoo/enterprise#40975

X-original-commit: 9bd26b234126ab2a07266e3aabc0723de423906f
Signed-off-by: William André (wan) <wan@odoo.com>