[FIX] account_report: Fixing way we hide/show lines in reports
Before this fix, we used .show() and .hide() to toggle lines'
display. These 2 functions add a style "display: block" and
"display: none" on the selected element. It caused a problem
because the table row should have a "display: table-row"
instead.
Now, we're toggling a specific class to hide lines or not.
This class adds a "display: none" on the line.

opw-3236881
opw-3237049

closes odoo/enterprise#40342

X-original-commit: a8bc5c6d940d37e22338e0788b26237a927e62a5
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>