[FIX] sale_amazon: prevent traceback on linking amazon account
Description
This PR fixes an issue where a traceback occurs while creating a new Amazon
account in the sale_amazon module.

Steps to Reproduce
- Install the `sale_amazon` module
- Create a new company and switch to it
- In the Sale app, go to Configuration, and then to the Connectors section.
- Click on Amazon Accounts.
- Click on "New", fill in the required information, and click on "Save".
-  A traceback will appear.

Cause
The traceback occurs because the "company" value is not present in the "vals"
dictionary. This is because we cannot access read-only fields during the create
method.

Solution
To fix this issue, we added `force_save` to the company_id field. This ensures that the "company" value is
available and the traceback is avoided.

sentry-4108896286

closes odoo/enterprise#40762

X-original-commit: 9aa426b564c6de7909c4d20d1ac12a7398daf7af
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>