[FIX] planning: select right employee to check the progressbar in tour
Before this commit, a step could fail in planning tour because the
progress selected to check if it is visible is the first one found
instead of selecting the one of the employee created in the test.
It means when there are some data (demo data for instance), the
progressbar checked if not the one of the employee created in the
test and so the step could fail because the progressbar is hidden
for the employee selected.

This commit selects progressbar of the employee created in the test
to be sure we select the progressbar of the right employee and to have
a consistent test.

closes odoo/enterprise#40406

X-original-commit: fe99f481ee8cac7e210af2e8103191df1ab63f53
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>