[FIX] website_helpdesk_form: correct ticket id on submitted page
Before this commit:
If a public users create two helpdesk ticket from the website form the
ticket id mentioned on the website form will always be the same.

As the page is cached, the function `_website_form_last_record` is not
called again when the page is opened which result on the same page being
render.
So any public users (even on different browser/sessions) will have the
same ticket ID mentioned which is wrong (as long as the page is cached).

After this commit:
The ticket given on the submitted page would correspond to the created
ticket

Note: the issue does not happen on version 16 (and >), probably as the
field `cached_time` have been removed, see:
https://github.com/odoo/odoo/commit/0452d0701fcb8f4588b9b05dad9bbdd507539265
It does not reproduce on version 15 neither as this issue is already
fixed in the same way, see:
https://github.com/odoo/enterprise/pull/29778

opw-3141250

closes odoo/enterprise#40690

X-original-commit: e7ed33ab46d4b9b5fec8e0c4a72e5e81a1d69273
Signed-off-by: Sens Loan (lse) <lse@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>