[IMP] account_accountant: bank recon label display
The liquidity line of the bank recon truncates the label.
With this task, the label column is removed, and is now placed on the next line.
To save space the strikethrough values for debit/credit/amount_currency (partial reconcile) - are moved to the same line as the label

Part of:
Task-3279342

closes odoo/enterprise#40867

X-original-commit: 5cb3a7d3b6e7f830bf6fc51157fcc4b714fa25e8
Signed-off-by: Laurent Smet <las@odoo.com>