[IMP] account_accountant: 16.0 bank recon label display
Backport and cherry-pick of:
5cb3a7d3b6e7f830bf6fc51157fcc4b714fa25e8
488e68210d63a4d63d52c962a8381b128ef7cc96

The liquidity line of the bank recon truncates the label.
With this task, the label column is removed, and is now placed on the next line.
To save space the strikethrough values for debit/credit/amount_currency (partial reconcile) - are moved to the same line as the label

Task-3414726
(original 3279342)

closes odoo/enterprise#44351

Signed-off-by: Laurent Smet (las) <las@odoo.com>