[IMP] {website_}delivery_sendcloud: Set Delivery By Product
A Sendcloud delivery carrier now relies on a Sendcloud's product
(that is a set of Sendcloud's methods) rather than on a single
Sendcloud's method.

As the parcel's shipping however still rely on a defined method, some
models were enhanced/redefined as some functions and checks were added
to handle the method selection.

Task #3109757

Upgrade : odoo/upgrade#5034

closes odoo/enterprise#41079

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>