[FIX] sign: don't show terms page in sitemap if not enabled
If the sign terms page is not enabled, it will return an "error" page:
while it will return a 200 HTTP code, it will display a message like
`Oops: The requested page is invalid` which doesn't bring much value in
your sitemap/indexed pages.

To be enabled, those two settings should be checked:
- `Sign Default Terms & Conditions` in sign settings
- `Terms as Web Page` as sub option

This commit is simply conditionning the presence of this page in the
sitemap so it's only shown when the terms page is actually enabled.

Note that when enabled, it's important to have it in the sitemap for SEO
as people might want to find your terms through google etc.

opw-3255831

closes odoo/enterprise#40945

X-original-commit: 8de129ba09a7b34a2d40712e26e03f9c3323bee0
Signed-off-by: Romain Derie (rde) <rde@odoo.com>