[IMP] documents: use separate props for selected record and inspected records
Purpose
=======
The current logic of DocumentInspector is super weird. When you want
to inspect a document, you need to select it.

But you can also inspect a document that is not selected, the current code
- save the current selection
- select the document we want to preview
- when we close the preview, it restores the selection

A better way to do it is to have two separate variables to differentiate
the selection and the documents previewed.

Task-3290796

Part-of: odoo/enterprise#40441