[FIX] helpdesk: remove "Assign to me" from helpdesk ticket email header
Current behaviour:
On a brand new helpdesk ticket with no assignees, when we add an user to follow
the ticket with an email, there is a link "Assign to me" that is supposed
to assign the newly following user as the assignee of the ticket. This
link is dead, returning a 404.

Expected behaviour:
Either correct the link, or remove it from the email header.

Steps to reproduce:
- Install Helpdesk
- Create a ticket, with no assignees
- Add a new follower to the task with Send Email checked.
- Check the outgoing email, and click on the link "Assign to me".
- 404 response

Reason for the problem:
The route which we were using for this `/mail/assign` doesn't exist
anymore since version 15. This was an intended design change, to
remove noice. The user can click on "View Helpdesk Ticket" and assign
himself on there.

Fix:
Remove the "Assign to me" link from the email header.
Same as in PR #119821.

Affected versions:
- 15.0
- saas-15.2
- 16.0
- saas-16.1
- saas-16.2
- master

opw-3245935

closes odoo/enterprise#40743

X-original-commit: 89510da63710ef3207246f3edc6989c5046d6879
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>