[FIX] l10n_co_edi: IMP_2 should be positive
There is not much documentation available, but a change in the community
code is making the test to fail.
That change fixes the fact that balance was not computed correctly if
only the amount in foreign currency is given, so that amount was 0 in
the test when it shouldn't have been.
Now that this is fixed, we see that it makes sense to take the absolute
value, just like when the tax code is `05`

closes odoo/enterprise#40467

X-original-commit: f6cfef5aa7de4b12d12139f5474a50daa5314980
Related: odoo/odoo#120022
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>