[FIX] knowledge: hide "New" button in readonly articles' embeds
Purpose:
--------
Currently, the embedded item views (list and kanban) show the "New"
button if the user does not have write permission on the active article.
Clicking on this button in that case results in an access error, since
the user tries to add a child to the current article, which he is not
allowed to do.
This commit hides the "New" button for these item views when the user
does not have write permission on the current active article, to prevent
triggering these access errors.

Task-3292052

closes odoo/enterprise#40432

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>