[FIX] knowledge: remove usage of container query (not supported on IOS)
Revert the changes introduced in
https://github.com/odoo/enterprise/pull/38827
related to the introduction of `@container` (container query), because we want
to support Safari on IOS version `>=14` and that feature is only available on
`>= 16`.

The usage of that container query also introduced some positioning bugs in
Kanban embedded views, because with a container, the position is relative to the
container and not the viewport. Removing that CSS features also fixe those bugs.

Rework the styling to improve responsiveness of Behavior buttons in a `/table`
(`/file` and `/template (clipboard)`).

Introduce the BehaviorToolbar component. As we already have 4 blocks in
knowledge using the same toolbar style, it becomes interesting to create a
generic component, so that style updates in the future are consistent across all
blocks.

Task-3291771

closes odoo/enterprise#42168

X-original-commit: aa902179c3a130bebc4bb3bf5caeff4f4322dded
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>